requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
python-dotenv
Pillow
openai
pydub
spotdl>=3.9.0
youtubesearchpython>=1.6.2
yt-dlp
google-generativeai
